EduBot/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SocketIO==5.3.6
Flask-Limiter==3.5.0
PyMySQL==1.1.0
cryptography==41.0.7
Werkzeug==2.3.7
WTForms==3.1.0
python-socketio==5.10.0
python-engineio>=4.8.0
eventlet==0.33.3
python-dotenv==1.0.0
email-validator==2.1.0
Pillow==10.1.0
google-generativeai==0.8.5
